pyproject.toml
[project]
name = "sec-analyzer"
version = "0.1.0"
description = "CLI tool for analysts to assess company risks from SEC reports using AI agents"
readme = "README.md"
authors = [
    { name = "Kaushik Ashodiya", email = "kaushik.ashodiya@company.org" }
]
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "strands-agents>=0.1.0",
    "boto3>=1.34.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "aiosqlite>=0.19.0",
    "edgartools>=2.0.0",
]

[project.scripts]
sec-analyzer = "sec_analyzer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sec_analyzer"]
